requirements.txt
pytorch
mmcv-full==1.3.11
mmdet==2.11.0
